Move ewelinkModernExtend to lib/ewelink
